election/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
Faker==13.3.1
PyMySQL==1.0.2
